style(msvc): fix C4244/C4267 in OptionsList/EditMenu/CubicSpline/TimingData/MusicWheel/CourseLoaderCRS + 3 residual sites (item 2)
All static_cast, documenting existing intentional narrowing -- no
logic changes. sm_tests unchanged at 5966/226; test_TimingData.cpp
and test_NotesLoaderCRS.cpp re-checked identical.

- OptionsList.cpp (10): unsigned/int locals and a GetScreen(int) call
  fed size_t choice counts/indices; two wrap(int&, size_t) sites; a
  linear-search loop returning its size_t index as int.
- EditMenu.cpp (10): two reverse-loop size()-1 bounds; six one-line
  GetRowSize accessors returning size() as int; a size()-1
  default-selection index.
- CubicSpline.cpp (10): a shared LCSN_EVAL_SOMETHING macro fixed once
  covers 4 call sites; three more lua_createtable/lua_rawgeti/
  lua_rawseti sites in the coefficient get/set helpers; two
  lua_pushnumber(size_t) accessors.
- TimingData.cpp (9, SS5-protected TimingData, characterization-only):
  a 4-way size() sum into an unsigned int segment count; an
  EraseSegment(int, ...) fed a size_t loop index at 2 sites; a
  binary-search size()-1 upper bound; a lua_createtable/lua_rawseti
  block serializing timing segments to Lua.
- MusicWheel.cpp (9): a MusicWheelItemData ctor's int section-count
  param fed size_t at 2 sites; two unsigned snapshot-size locals; a
  std::min<unsigned int> call's second arg; two wrap(int&, size_t)
  sites; a reverse-loop bound; a RandomInt(size_t) call.
- CourseLoaderCRS.cpp (9, SS5-protected .crs, characterization-only):
  four Left/Right pairs fed strlen()/size()-derived lengths parsing
  BEST/WORST/GRADEBEST/GRADEWORST song-choice prefixes; an int
  (StringToInt) assigned into a float field.
- Residual C4267 the broken C424[47] regex had hidden in 3 files
  already marked done earlier this sweep: NoteField.cpp (5: two int
  column-index members fed a size_t loop var, a binary-search
  size()-1 bound, a lua_createtable/lua_rawseti pair pushing
  column-renderer actors); ScreenGameplay.cpp (2: an unsigned int
  queue-size local, a reverse-loop size()-1 bound);
  ScreenOptionsMasterPrefs.cpp (1: a MoveMap(..., unsigned cnt) call
  fed size_t).

+++ src/CourseLoaderCRS.cpp
@@ -392,9 +392,9 @@ bool CourseLoaderCRS::ParseCourseSong( const MsdFile::value_t &sParams, CourseEn
 	// to a lack of songs. -aj
 	int iNumSongs = SONGMAN->GetNumSongs();
 	// most played
-	if( sParams[1].Left(strlen("BEST")) == "BEST" )
+	if( sParams[1].Left(static_cast<int>(strlen("BEST"))) == "BEST" )
 	{
-		int iChooseIndex = StringToInt( sParams[1].Right(sParams[1].size()-strlen("BEST")) ) - 1;
+		int iChooseIndex = StringToInt( sParams[1].Right(static_cast<int>(sParams[1].size()-strlen("BEST"))) ) - 1;
 		if( iChooseIndex > iNumSongs )
 		{
 			// looking up a song that doesn't exist.
@@ -408,9 +408,9 @@ bool CourseLoaderCRS::ParseCourseSong( const MsdFile::value_t &sParams, CourseEn
 		new_entry.songSort = SongSort_MostPlays;
 	}
 	// least played
-	else if( sParams[1].Left(strlen("WORST")) == "WORST" )
+	else if( sParams[1].Left(static_cast<int>(strlen("WORST"))) == "WORST" )
 	{
-		int iChooseIndex = StringToInt( sParams[1].Right(sParams[1].size()-strlen("WORST")) ) - 1;
+		int iChooseIndex = StringToInt( sParams[1].Right(static_cast<int>(sParams[1].size()-strlen("WORST"))) ) - 1;
 		if( iChooseIndex > iNumSongs )
 		{
 			// looking up a song that doesn't exist.
@@ -424,16 +424,16 @@ bool CourseLoaderCRS::ParseCourseSong( const MsdFile::value_t &sParams, CourseEn
 		new_entry.songSort = SongSort_FewestPlays;
 	}
 	// best grades
-	else if( sParams[1].Left(strlen("GRADEBEST")) == "GRADEBEST" )
+	else if( sParams[1].Left(static_cast<int>(strlen("GRADEBEST"))) == "GRADEBEST" )
 	{
-		new_entry.iChooseIndex = StringToInt( sParams[1].Right(sParams[1].size()-strlen("GRADEBEST")) ) - 1;
+		new_entry.iChooseIndex = StringToInt( sParams[1].Right(static_cast<int>(sParams[1].size()-strlen("GRADEBEST"))) ) - 1;
 		CLAMP( new_entry.iChooseIndex, 0, 500 );
 		new_entry.songSort = SongSort_TopGrades;
 	}
 	// worst grades
-	else if( sParams[1].Left(strlen("GRADEWORST")) == "GRADEWORST" )
+	else if( sParams[1].Left(static_cast<int>(strlen("GRADEWORST"))) == "GRADEWORST" )
 	{
-		new_entry.iChooseIndex = StringToInt( sParams[1].Right(sParams[1].size()-strlen("GRADEWORST")) ) - 1;
+		new_entry.iChooseIndex = StringToInt( sParams[1].Right(static_cast<int>(sParams[1].size()-strlen("GRADEWORST"))) ) - 1;
 		CLAMP( new_entry.iChooseIndex, 0, 500 );
 		new_entry.songSort = SongSort_LowestGrades;
 	}
@@ -669,7 +669,7 @@ bool CourseLoaderCRS::ParseCourseSongSelect(const MsdFile::value_t &sParams, Cou
 		}
 		else if( sParamName.EqualsNoCase("GAINSECONDS") )
 		{
-			new_entry.fGainSeconds = StringToInt(sParamValue);
+			new_entry.fGainSeconds = static_cast<float>(StringToInt(sParamValue));
 		}
 		else if( sParamName.EqualsNoCase("MODS") )
 		{
